Move test over to samples
