+++ libretro-common/utils/md5.c
@@ -293,54 +293,3 @@ void MD5_Final(unsigned char *result, MD5_CTX *ctx)
 
 	memset(ctx, 0, sizeof(*ctx));
 }
-
-#ifdef MD5_BUILD_UTILITY
-
-#include <stdio.h>
-
-int main (int argc, char *argv[])
-{
-   /* For each command line argument in turn:
-    ** filename          -- prints message digest and name of file
-    */
-   int i;
-   MD5_CTX ctx;
-   FILE* file;
-   size_t numread;
-   char buffer[16384];
-   unsigned char result[16];
-
-   for (i = 1; i < argc; i++)
-   {
-      MD5_Init(&ctx);
-      file = fopen(argv[i], "rb");
-
-      if (file)
-      {
-         do
-         {
-            numread = fread((void*)buffer, 1, sizeof(buffer), file);
-
-            if (numread)
-            {
-               MD5_Update(&ctx,(void*)buffer, numread);
-            }
-         }
-         while (numread);
-
-         fclose(file);
-         MD5_Final(result, &ctx);
-         printf("%02x%02x%02x%02x%02x%02x%02x%02x"
-			          "%02x%02x%02x%02x%02x%02x%02x%02x %s\n",
-			          result[ 0 ], result[ 1 ], result[ 2 ], result[ 3 ],
-                result[ 4 ], result[ 5 ], result[ 6 ], result[ 7 ],
-                result[ 8 ], result[ 9 ], result[ 10 ], result[ 11 ],
-                result[ 12 ], result[ 13 ], result[ 14 ], result[ 15 ],
-                argv[i]);
-      }
-   }
-
-   return 0;
-}
-
-#endif
